BREAKING(gaussian): depricate gaussian::{val,fit}
